requirements.txt
streamlit
python-dotenv
groq
# Language Model & Chain Management
langchain
langchain-groq
langchain-community
# HTTP & GitLab API
requests
gitpython
